requirements.txt
streamlit
pandas
numpy
requests
plotly
Pillow